+++ database/migrations/2020_04_10_224546_update_short_url_table_for_version_three_zero_zero.php
@@ -1,0 +1,33 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+class UpdateShortURLTableForVersionThreeZeroZero extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::table('short_urls', function (Blueprint $table) {
+            $table->timestamp('activated_at')->after('track_device_type')->nullable()->default(now());
+            $table->timestamp('deactivated_at')->after('activated_at')->nullable();
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::table('short_urls', function (Blueprint $table) {
+            $table->dropColumn(['activated_at', 'deactivated_at']);
+        });
+    }
+}